[Refactor] Move AnalysisResult into an own file
