+++ lib/cocoapods/installer/analyzer.rb
@@ -6,8 +6,8 @@ module Pod
     class Analyzer
       include Config::Mixin
 
-      autoload :SandboxAnalyzer, 'cocoapods/installer/analyzer/sandbox_analyzer'
-
+      autoload :AnalysisResult,            'cocoapods/installer/analyzer/analysis_result'
+      autoload :SandboxAnalyzer,           'cocoapods/installer/analyzer/sandbox_analyzer'
       autoload :LockingDependencyAnalyzer, 'cocoapods/installer/analyzer/locking_dependency_analyzer'
 
       # @return [Sandbox] The sandbox where the Pods should be installed.
@@ -764,45 +764,6 @@ module Pod
                 raise Informative, 'It is necessary to specify the platform in the Podfile if not integrating.'
               end
             end
-          end
-        end
-      end
-
-      #-----------------------------------------------------------------------#
-
-      class AnalysisResult
-        # @return [SpecsState] the states of the Podfile specs.
-        #
-        attr_accessor :podfile_state
-
-        # @return [Hash{TargetDefinition => Array<Spec>}] the specifications
-        #         grouped by target.
-        #
-        attr_accessor :specs_by_target
-
-        # @return [Array<Specification>] the specifications of the resolved
-        #         version of Pods that should be installed.
-        #
-        attr_accessor :specifications
-
-        # @return [SpecsState] the states of the {Sandbox} respect the resolved
-        #         specifications.
-        #
-        attr_accessor :sandbox_state
-
-        # @return [Array<Target>] The Podfile targets containing library
-        #         dependencies.
-        #
-        attr_accessor :targets
-
-        # @return [Hash{String=>Symbol}] A hash representing all the user build
-        #         configurations across all integration targets. Each key
-        #         corresponds to the name of a configuration and its value to
-        #         its type (`:debug` or `:release`).
-        #
-        def all_user_build_configurations
-          targets.reduce({}) do |result, target|
-            result.merge(target.user_build_configurations)
           end
         end
       end
